+++ packages/dddart/lib/dddart.dart
@@ -61,7 +61,6 @@ export 'src/entity.dart';
 export 'src/event_bus.dart';
 export 'src/file_log_handler.dart';
 export 'src/in_memory_repository.dart';
-export 'src/queryable_repository.dart';
 export 'src/repository.dart';
 export 'src/repository_exception.dart';
 export 'src/uuid_value.dart';